feat: intraday price snapshots, job tracking, news datetime, IR calendar
- Add price_snapshots table for hourly intraday price collection (MOEX ISS)
- Scheduler runs fetch_snapshots every hour 10:00-19:00 MSK on weekdays
- Auto-cleanup of snapshots older than 30 days
- Add JobRun model for tracking all scheduler/manual job executions
- Change news.date from Date to DateTime for freshness checks
- Add IR calendar fetcher (MOEX ISS corporate events → catalysts)
- Add fetch-snapshots and fetch-ir-calendar API triggers
- Update NEWS_REACTION_GUIDE with intraday snapshot usage
- Fix shares_out type: BigInteger → Numeric(14,2)
- Add GZip middleware, admin panel improvements

+++ investment-api/app/schemas/company.py
@@ -22,7 +22,7 @@ class CompanyCreate(BaseModel):
     current_price: Optional[Decimal] = None
     upside: Optional[Decimal] = None
     market_cap: Optional[Decimal] = None
-    shares_out: Optional[int] = None
+    shares_out: Optional[Decimal] = None
     free_float: Optional[Decimal] = None
     adv_rub_mln: Optional[Decimal] = None
     p_e: Optional[Decimal] = None
@@ -42,7 +42,7 @@ class CompanyUpdate(BaseModel):
     current_price: Optional[Decimal] = None
     upside: Optional[Decimal] = None
     market_cap: Optional[Decimal] = None
-    shares_out: Optional[int] = None
+    shares_out: Optional[Decimal] = None
     free_float: Optional[Decimal] = None
     adv_rub_mln: Optional[Decimal] = None
     p_e: Optional[Decimal] = None
@@ -87,7 +87,7 @@ class CompanyResponse(BaseModel):
     current_price: Optional[Decimal] = None
     upside: Optional[Decimal] = None
     market_cap: Optional[Decimal] = None
-    shares_out: Optional[int] = None
+    shares_out: Optional[Decimal] = None
     free_float: Optional[Decimal] = None
     adv_rub_mln: Optional[Decimal] = None
     p_e: Optional[Decimal] = None
